# Try out static instance tracer field [DO NOT MERGE]