@creamsoup @ejona86 @zhangkun83 Looking for your thoughts on this


Trying out a new API for perfmark.   This makes the PerfMark class not a collection of static methods, but a singleton holder of itself. There are some benefits to this:

1.  The `Impl` class is gone, shrinking the api jar and method count.   Looks like maybe 1KB.
2.  It is much harder for outside classes to extend `Impl`.  I originally didnt want to leak the unpacking / packing methods in Impl to the out side, but there wasnt a good way to do so.   Making those methods static would have made it possible for outsiders to subclass it and call the static methods directly.  PerfMark and Impl are rolled in together now.
3.  It is now impossible for subclasses of Impl (now PerfMark) to be instantiated.  The ctor now ensures that only clinit can construct the instance, 
4.  Access is now through the `tracer()` static method.  I kind of like this because it can be statically imported, without having to litter the class name `PerfMark` everywhere.

In terms of making this work with the existing classes without break api, I was thinking of maybe making a sep package called `perfmark-tracer` at `io.perfmark.trace`, and maybe naming the class Tracer.    The real impl would like inside perfmark-impl just like API does, so existing calls sites would still work.

Thoughts?