@carl-mastrangelo 2019-10-24
Some initial benchmarks:

```
Before:
Benchmark                   (enabled)  Mode  Cnt   Score   Error  Units
EnabledBenchmark.createTag       true  avgt    5   5.512 ± 0.230  ns/op
EnabledBenchmark.createTag      false  avgt    5   2.714 ± 0.018  ns/op
EnabledBenchmark.event           true  avgt    5  32.612 ± 0.559  ns/op
EnabledBenchmark.event          false  avgt    5   0.610 ± 0.008  ns/op
EnabledBenchmark.link            true  avgt    5  12.487 ± 0.200  ns/op
EnabledBenchmark.link           false  avgt    5   0.602 ± 0.005  ns/op
EnabledBenchmark.startStop       true  avgt    5  60.258 ± 1.927  ns/op
EnabledBenchmark.startStop      false  avgt    5   0.601 ± 0.007  ns/op

After:
Benchmark                   (enabled)  Mode  Cnt   Score   Error  Units
EnabledBenchmark.createTag       true  avgt    5   5.695 ± 0.548  ns/op
EnabledBenchmark.createTag      false  avgt    5   3.000 ± 0.045  ns/op
EnabledBenchmark.event           true  avgt    5  34.924 ± 1.461  ns/op
EnabledBenchmark.event          false  avgt    5   0.619 ± 0.005  ns/op
EnabledBenchmark.link            true  avgt    5  12.901 ± 0.371  ns/op
EnabledBenchmark.link           false  avgt    5   0.629 ± 0.004  ns/op
EnabledBenchmark.startStop       true  avgt    5  64.067 ± 0.757  ns/op
EnabledBenchmark.startStop      false  avgt    5   0.601 ± 0.011  ns/op

```