@ejona86 2019-11-01
We are reading the benchmarks as "no change?" The 4µs increase is noticable, but there shouldn't be any difference so I'm okay if we're claiming that is noise.

1. We don't expect any improvement for Android, though, right? Since things should already be inlined. Or were the methods too large to be inlined by default?
2. This doesn't seem so much about extending but hiding internals. The new approach is now leaking the internals to users more directly as noise, as they will be using the PerfMark class already and protected methods are displayed lots of places. Maybe we can put them as static methods in a protected PerfMark.Impl class? (Or call it Unsafe to mirror Netty; whatever.)
3. I'm not quite sure I see the value. But it doesn't seem to hurt anything. It seems this was mostly the case already, except for reflection or doing a hack of creating a class in the io.perfmark package.
4. I don't see how `PerfMark.startTask();` is any more littered than `tracer().startTask();`. It's even same number of letters and same number of shift-keys. I prefer PerfMark since it is much more informative when reading (there's lots of different types of tracers). But I'm not bothered with doing `PerfMark.tracer()...`; I just don't see the claimed advantage

Concerning `io.perfmark.trace`, are you saying that there would be another artifact for this new API and we'd stop using the current one? If `Tracer` is the new name then code would always be encouraged to have the static import for `tracer()`... That seems too generic of a name to be considered a net win. But we could bikeshed over that.

So (2) I understand, but the other benefits I don't see. I see the benefit in (2), but I'm not sure it is worth the change. I don't feel the need to bend over backwards to save a user from themselves and I don't feel like this will provide any real benefit to users. I can't figure out what someone would gain by extending Impl to get at the pack/unpack. If that's a strong concern, I might want to think more about alternatives, but I do accept that many of the normal tricks incur costs.