@carl-mastrangelo 2019-11-09
@ejona86 thanks for the reply.  I am probably going to abandon this, the original reason was to avoid the static function call and avoid one level of inlining.   (In addition to shrinking binary size for android).