@ejona86 2019-11-11
Oh, one level of inlining... That would be an additional benefit.

What is the problem with a static function call? Isn't the only problem the visibility of the method?