@carl-mastrangelo 2019-11-11
@ejona86 There's another growing concern that Is hard to work into the static function model:  working with java 8 types for Lambdas.   I have run into problems where I would really like to defer calculation of the Tag, or Task name, and the current Java6 targeted API makes that hard.    I wanted to be able to expose tracing calls that accept lambdas.    If the tracer was an object, it could be cast to a subtype that is java8 specific and expose the methods, leaving `PerfMark` as the super class and only having the Java6 methods.


Having a separate top level tracer type would be a little ugly IMO, (such as `PerfMark8`) with all the static methods, and would look ugly.      I'm trying to figure out a way to expose the advanced tracer calls without breaking Android.  Any ideas here would be welcome.

One more thing:   Android does allow  desugaring lambdas, which means PerfMark could support it, but Android /does not/ allow using any of the library types (like `Supplier`).    PerfMark can bundle this type to allow callers to pass in lambdas, but it would bloat the binary, and not be very java 8 idiomatic.   Again, Looking for ideas on how to support Java8 without hurting android.